Revert "add sale"
This reverts commit e9a930efee24f62c6633222c4680879b12df937a.

+++ PdH.Business/Interfaces/IProductBc.cs
@@ -35,9 +35,9 @@ namespace PdH.Business.Interfaces
 
         Product ChangeStatus(Product product);
 
-        Product AddStock(Product product, long stock);
+        Product AddStock(Product product, int stock);
 
-        Product RemoveStock(Product product, long sellQuantity);
+        Product RemoveStock(Product product, int sellQuantity);
 
         void Delete(Product product);
